extract module for http-client
